+++ Frameworks/AddressesKit/Frameworks/AddressView/ADPersonPropertyView.m
@@ -315,7 +315,8 @@ static float _globalFontSize;
 
 - (BOOL) updatePersonWithMultiValueFromCell: (ADPersonPropertyCell*) cell
 {
-  NSString *key, *label, *identifier; id value; int i;
+  NSString *key, *label, *identifier; id value;
+  NSInteger i;
   ADPropertyType type; ADMutableMultiValue *mv;
   NSMutableDictionary *dict;
 
@@ -627,9 +628,9 @@ static float _globalFontSize;
   [self beginEditingInCellAtIndex: [_cells count]-1 countingBackwards: YES];
 }
 
-- (int) indexOfEditableCellWithDetails: (id) details
+- (NSInteger) indexOfEditableCellWithDetails: (id) details
 {
-  int i;
+  NSInteger i;
 
   for(i=0; i<[_cells count]; i++)
     {
@@ -648,7 +649,7 @@ static float _globalFontSize;
 
 - (NSString*) propertyForDragWithDetails: (id) details
 {
-  NSString *identifier; int index; ADMultiValue *mv;
+  NSString *identifier; NSInteger index; ADMultiValue *mv;
   id value;
 
   if(!details &&
@@ -828,7 +829,7 @@ static float _globalFontSize;
 - (void) textDidEndEditing: (NSNotification*) aNotification
 {
   int textMovement = -1;
-  int cellIndex;
+  NSInteger cellIndex;
   id c;
   NSDictionary *dict = [aNotification userInfo];
   id details;
